pyproject.toml
[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.12.0"
aiohttp = "^3.8.0"
beautifulsoup4 = "^4.9.3"
duckduckgo-search = "^3.0.0"
lxml = {extras = ["html_clean"], version = "^5.1.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pymilvus = "^2.4.9"